# docs/swagger/ai-service.json
{"openapi":"3.1.0","info":{"title":"Moni AI Service API","description":"기업 이슈 분석 및 뉴스 수집 API","version":"v1.0.0"},"servers":[{"url":"http://172.18.0.9:19096","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/api/v1/ai/{ticker}/issue-analysis":{"get":{"tags":["ai-controller"],"summary":"기업 이슈 분석 조회","description":"캐시된 분석 결과를 조회합니다. LLM 호출 없이 저장된 결과를 반환합니다.","operationId":"getIssueAnalysis","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GlobalResponseCompanyIssueResDto"}}}},"404":{"description":"분석 결과 없음 또는 등록되지 않은 ticker","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GlobalResponseCompanyIssueResDto"}}}}}},"post":{"tags":["ai-controller"],"summary":"기업 이슈 분석 생성","description":"RAG 기반으로 기업 이슈를 분석합니다. 캐시가 있을 경우 400을 반환합니다.","operationId":"createIssueAnalysis","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"201":{"description":"분석 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GlobalResponseCompanyIssueResDto"}}}},"400":{"description":"이미 분석된 기업","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GlobalResponseCompanyIssueResDto"}}}},"404":{"description":"등록되지 않은 ticker","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GlobalResponseCompanyIssueResDto"}}}},"500":{"description":"LLM 호출 실패","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GlobalResponseCompanyIssueResDto"}}}}}}},"/api/v1/ai/portfolio/analysis":{"post":{"tags":["portfolio-analysis-controller"],"operationId":"analyze","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PortfolioAnalysisRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GlobalResponsePortfolioAnalysisResponseDto"}}}}}}},"/api/v1/ai/news-summary":{"post":{"tags":["ai-controller"],"summary":"시장 뉴스 분석 생성","description":"키워드 기반 시장 뉴스를 분석합니다.","operationId":"createNewsAnalysis","parameters":[{"name":"keyword","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"201":{"description":"분석 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GlobalResponseMarketAnalysisResDto"}}}},"404":{"description":"등록되지 않은 키워드","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GlobalResponseMarketAnalysisResDto"}}}},"500":{"description":"LLM 호출 실패","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GlobalResponseMarketAnalysisResDto"}}}}}}},"/api/v1/admin/ai/news/ticker":{"post":{"tags":["news-admin-controller"],"summary":"뉴스 직접 등록","description":"ticker와 회사명 검증 후 뉴스를 직접 등록하고 벡터 저장합니다.","operationId":"createNews","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewsCreateReqDto"}}},"required":true},"responses":{"201":{"description":"등록 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GlobalResponseNewsCreateResDto"}}}},"400":{"description":"ticker와 회사명 불일치","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GlobalResponseNewsCreateResDto"}}}},"404":{"description":"등록되지 않은 ticker","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GlobalResponseNewsCreateResDto"}}}},"409":{"description":"이미 등록된 뉴스","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GlobalResponseNewsCreateResDto"}}}}}}},"/api/v1/admin/ai/news/market/fetch":{"post":{"tags":["news-admin-controller"],"summary":"시장 뉴스 수동 수집","description":"코스피, 금리, 달러 등 시장 전반 뉴스를 수동으로 수집합니다.","operationId":"fetchMarketNews","responses":{"201":{"description":"수집 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GlobalResponseString"}}}},"500":{"description":"수집 실패","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GlobalResponseString"}}}}}}},"/api/v1/admin/ai/news/fetch":{"post":{"tags":["news-admin-controller"],"summary":"기업 뉴스 수동 수집","description":"등록된 전체 기업의 뉴스를 수동으로 수집합니다.","operationId":"createNewsVector","responses":{"201":{"description":"수집 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GlobalResponseString"}}}},"500":{"description":"수집 실패","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GlobalResponseString"}}}}}}},"/api/v1/ai":{"get":{"tags":["ai-controller"],"summary":"관심 기업 목록 조회","description":"뉴스 수집 대상 기업 목록을 반환합니다.","operationId":"getWatchCompany","responses":{"200":{"description":"조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GlobalResponseListWatchCompanyResDto"}}}}}}},"/api/v1/ai/news":{"get":{"tags":["ai-controller"],"summary":"뉴스 목록 조회","description":"ticker, companyName, keyword로 필터링, 날짜 기준 최대 3일 조회","operationId":"getNewsList","parameters":[{"name":"request","in":"query","required":true,"schema":{"$ref":"#/components/schemas/NewsSearchReqDto"}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GlobalResponsePageResNewsResDto"}}}}}}}},"components":{"schemas":{"CompanyIssueResDto":{"type":"object","properties":{"ticker":{"type":"string"},"company_name":{"type":"string"},"summary":{"type":"string"},"sentiment":{"type":"string","enum":["POSITIVE","NEGATIVE","NEUTRAL"]},"analyzedAt":{"type":"string","format":"date-time"},"expiredAt":{"type":"string","format":"date-time"}}},"ErrorResponse":{"type":"object","properties":{"errorClassName":{"type":"string"},"message":{"type":"string"}}},"GlobalResponseCompanyIssueResDto":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/CompanyIssueResDto"},"errors":{"$ref":"#/components/schemas/ErrorResponse"}}},"PortfolioAnalysisRequestDto":{"type":"object","properties":{"analysisId":{"type":"string","format":"uuid"},"userId":{"type":"string","format":"uuid"},"totalEvaluationAmount":{"type":"number","minimum":0.00},"totalReturnRate":{"type":"number"},"concentrationScore":{"type":"number","minimum":0.00},"concentrationThreshold":{"type":"number","minimum":0.00},"holdings":{"type":"array","items":{"$ref":"#/components/schemas/PortfolioHoldingRequestDto"},"minItems":1},"tendencyAnalysis":{"$ref":"#/components/schemas/PortfolioTendencyAnalysisRequestDto"}},"required":["analysisId","concentrationScore","concentrationThreshold","holdings","totalEvaluationAmount","totalReturnRate","userId"]},"PortfolioHoldingRequestDto":{"type":"object","properties":{"ticker":{"type":"string","minLength":1},"stockName":{"type":"string","minLength":1},"quantity":{"type":"integer","format":"int64"},"averagePurchasePrice":{"type":"number","minimum":0.00},"currentPrice":{"type":"number","minimum":0.00},"evaluationAmount":{"type":"number","minimum":0.00},"profitLoss":{"type":"number"},"profitRate":{"type":"number"},"weight":{"type":"number","maximum":100.00,"minimum":0.00}},"required":["averagePurchasePrice","currentPrice","evaluationAmount","profitLoss","profitRate","quantity","stockName","ticker","weight"]},"PortfolioTendencyAnalysisRequestDto":{"type":"object","properties":{"userTendencyType":{"type":"string","minLength":1},"userTendencyLabel":{"type":"string","minLength":1},"userTendencyScore":{"type":"integer","format":"int32","maximum":100,"minimum":0},"portfolioRiskType":{"type":"string","minLength":1},"portfolioRiskLabel":{"type":"string","minLength":1},"portfolioRiskScore":{"type":"integer","format":"int32","maximum":100,"minimum":0},"suitabilityScore":{"type":"integer","format":"int32","maximum":100,"minimum":0},"suitabilityLevel":{"type":"string","minLength":1}},"required":["portfolioRiskLabel","portfolioRiskScore","portfolioRiskType","suitabilityLevel","suitabilityScore","userTendencyLabel","userTendencyScore","userTendencyType"]},"GlobalResponsePortfolioAnalysisResponseDto":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/PortfolioAnalysisResponseDto"},"errors":{"$ref":"#/components/schemas/ErrorResponse"}}},"PortfolioAnalysisResponseDto":{"type":"object","properties":{"analysisId":{"type":"string","format":"uuid"},"summary":{"type":"string"},"tendencyAnalysis":{"$ref":"#/components/schemas/PortfolioTendencyAnalysisResponseDto"},"recommendation":{"type":"string"}}},"PortfolioTendencyAnalysisResponseDto":{"type":"object","properties":{"userTendencyType":{"type":"string"},"userTendencyLabel":{"type":"string"},"userTendencyScore":{"type":"integer","format":"int32"},"portfolioRiskType":{"type":"string"},"portfolioRiskLabel":{"type":"string"},"portfolioRiskScore":{"type":"integer","format":"int32"},"suitabilityScore":{"type":"integer","format":"int32"},"suitabilityLevel":{"type":"string"},"summary":{"type":"string"},"recommendation":{"type":"string"}}},"GlobalResponseMarketAnalysisResDto":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/MarketAnalysisResDto"},"errors":{"$ref":"#/components/schemas/ErrorResponse"}}},"MarketAnalysisResDto":{"type":"object","properties":{"summary":{"type":"string"}}},"NewsCreateReqDto":{"type":"object","properties":{"ticker":{"type":"string","maxLength":6,"minLength":0},"title":{"type":"string","minLength":1},"company_name":{"type":"string","minLength":1},"content":{"type":"string","minLength":1},"source":{"type":"string","maxLength":50,"minLength":0},"url":{"type":"string","minLength":1},"published_at":{"type":"string","format":"date-time"}},"required":["company_name","content","published_at","source","ticker","title","url"]},"GlobalResponseNewsCreateResDto":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/NewsCreateResDto"},"errors":{"$ref":"#/components/schemas/ErrorResponse"}}},"NewsCreateResDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"ticker":{"type":"string"},"title":{"type":"string"},"company_name":{"type":"string"},"content":{"type":"string"},"source":{"type":"string"},"url":{"type":"string"},"published_at":{"type":"string","format":"date-time"},"created_at":{"type":"string","format":"date-time"}}},"GlobalResponseString":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"string"},"errors":{"$ref":"#/components/schemas/ErrorResponse"}}},"GlobalResponseListWatchCompanyResDto":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/WatchCompanyResDto"}},"errors":{"$ref":"#/components/schemas/ErrorResponse"}}},"WatchCompanyResDto":{"type":"object","properties":{"ticker":{"type":"string"},"company_name":{"type":"string"}}},"NewsSearchReqDto":{"type":"object","properties":{"ticker":{"type":"string","maxLength":6,"minLength":0},"companyName":{"type":"string","maxLength":50,"minLength":0},"keyword":{"type":"string","maxLength":50,"minLength":0},"date":{"type":"string","format":"date"}}},"Pageable":{"type":"object","properties":{"page":{"type":"integer","format":"int32","minimum":0},"size":{"type":"integer","format":"int32","minimum":1},"sort":{"type":"array","items":{"type":"string"}}}},"GlobalResponsePageResNewsResDto":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/PageResNewsResDto"},"errors":{"$ref":"#/components/schemas/ErrorResponse"}}},"NewsResDto":{"type":"object","properties":{"title":{"type":"string"},"published_at":{"type":"string","format":"date-time"},"source":{"type":"string"}}},"PageResNewsResDto":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/NewsResDto"}},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"last":{"type":"boolean"}}}},"securitySchemes":{"bearerAuth":{"type":"http","name":"bearerAuth","scheme":"bearer","bearerFormat":"JWT"}}}}